asdf

+++ models.py
@@ -78,7 +78,7 @@ def rotation_model(inputs, train=True, norm=True, **kwargs):
     batch_size = inputs['images'].get_shape().as_list()[0]
 
     # rotations
-    rotation_labels = np.random.randint(0, 4, batch_size)
+    rotation_labels = np.random.randint(0, 4, batch_size,dtype='int32')
     input_to_network = tf.map_fn(lambda x: tf.image.rot90(x[0], x[1]), (inputs['images'], rotation_labels))
     outputs['labels_rotation'] = rotation_labels
     ### YOUR CODE HERE
